Basic Tracing
This commit uses AsyncLocalStorage to have some sort of ctx
that is shared and maintained throughout the async lifecycle
without us having to manually do such things. This allows us to
set some trace/span IDs and use that to trace functions and
flows throughout the system.

This allows for the end user to send `X-Trace-Id` headers and
that trace ID will be appended to each and every log, along with
the service itself breaking things up into logical spans